add delete cascade on q_a
Partway through adding second trigger (check_due) which determines if a test is taken after the due date. There's a function not connected to a trigger called set_due that is called as a default in test now, should not affect functionality. get_timestamp() is currently not broken but not functional either.